+++ src/server/server.h
@@ -684,10 +684,6 @@ qbool
 SV_ExecuteClientCommand(client_t *cl, const qchar *s);
 void
 SV_ClientThink(client_t *cl, usercmd_t *cmd);
-#if !defined(UDP_DOWNLOAD_OPTIMIZE)
-const qbool
-SV_WriteDownloadToClient(client_t *cl , msg_t *msg);
-#endif
 const qint
 SV_SendDownloadMessages(void);
 const qint